+++ app/src/main/java/com/eatssu/android/adapter/SnackAdapter.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.ui.menuAdapter
+package com.eatssu.android.adapter
 
 import android.content.Intent
 import android.view.LayoutInflater
@@ -6,14 +6,13 @@ import android.view.ViewGroup
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.eatssu.android.data.model.response.GetMenuInfoListResponse
-import com.eatssu.android.databinding.ItemDodamBinding
+import com.eatssu.android.databinding.ItemSnackBinding
 import com.eatssu.android.ui.review.ReviewListActivity
 
+class SnackAdapter(private val dataList: List<GetMenuInfoListResponse.MenuInfo>) :
+    RecyclerView.Adapter<SnackAdapter.ViewHolder>() {
 
-class DodamAdapter(private val dataList: List<GetMenuInfoListResponse.MenuInfo>):
-    RecyclerView.Adapter<DodamAdapter.ViewHolder>() {
-
-    inner class ViewHolder(private val binding: ItemDodamBinding) :
+    inner class ViewHolder(private val binding: ItemSnackBinding) :
         RecyclerView.ViewHolder(binding.root) {
 
         fun bind(position: Int) {
@@ -25,7 +24,7 @@ class DodamAdapter(private val dataList: List<GetMenuInfoListResponse.MenuInfo>)
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val binding =
-            ItemDodamBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+            ItemSnackBinding.inflate(LayoutInflater.from(parent.context), parent, false)
         return ViewHolder(binding)
     }
