+++ app/src/main/java/com/eatssu/android/ui/main/DinnerFragment.kt
@@ -9,13 +9,14 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.eatssu.android.adapter.DodamAdapter
+import com.eatssu.android.adapter.GisikAdapter
 import com.eatssu.android.data.enums.RestaurantType
 import com.eatssu.android.data.RetrofitImpl
 import com.eatssu.android.data.model.response.GetMenuInfoListResponse
 import com.eatssu.android.data.service.MenuService
 import com.eatssu.android.databinding.FragmentDinnerBinding
 import com.eatssu.android.ui.infopage.*
-import com.eatssu.android.ui.menuAdapter.*
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
